requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
Jinja2==3.1.5
python-frontmatter==1.1.0
bleach==6.1.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.0
Werkzeug==2.1.2
python-multipart
click==8.1.7
psutil==6.1.1
pyftpdlib==1.5.9
asyncio==3.4.3
gunicorn==21.2.0